some features added/changed
Changed the lock orientation so that if its changed in start/dead screen it wil change the field as well. An effect of this is that the matrix is not loaded when the document is ready but after the button is clicked.
now knows when there is no room left for new food.